# backend/docs/openapi.yaml
components:
  schemas:
...
    Gender:
      description: 性別
      enum:
      - male
      - female
      - other
      title: Gender
      type: string
    HTTPAuthorizationCredentials:
      description: 'The HTTP authorization credentials in the result of using `HTTPBearer`
        or

        `HTTPDigest` in a dependency.


        The HTTP authorization header value is split by the first space.


        The first part is the `scheme`, the second part is the `credentials`.


        For example, in an HTTP Bearer token scheme, the client will send a header

        like:


        ```

        Authorization: Bearer deadbeef12346

        ```


        In this case:


        * `scheme` will have the value `"Bearer"`

        * `credentials` will have the value `"deadbeef12346"`'
      properties:
        credentials:
          title: Credentials
          type: string
        scheme:
          title: Scheme
          type: string
      required:
      - scheme
      - credentials
      title: HTTPAuthorizationCredentials
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PasswordReset:
      description: パスワードリセットリクエスト
      properties:
        email:
          description: メールアドレス
          title: Email
          type: string
      required:
      - email
      title: PasswordReset
      type: object
...
    TokenResponse:
      description: トークンレスポンス
      properties:
        access_token:
          description: アクセストークン
          title: Access Token
          type: string
        expires_in:
          description: 有効期限（秒）
          title: Expires In
          type: integer
        token_type:
          default: bearer
          description: トークンタイプ
          title: Token Type
          type: string
        user:
          allOf:
          - $ref: '#/components/schemas/UserResponse'
          description: ユーザー情報
      required:
      - access_token
      - expires_in
      - user
      title: TokenResponse
      type: object
    UserCreate:
      description: ユーザー作成リクエスト
      properties:
        age:
          anyOf:
          - type: integer
          - type: 'null'
          description: 年齢
          title: Age
        email:
          description: メールアドレス
          title: Email
          type: string
        full_name:
          description: フルネーム
          title: Full Name
          type: string
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
          description: 性別
        height_cm:
          anyOf:
          - type: number
          - type: 'null'
          description: 身長（cm）
          title: Height Cm
        password:
          description: パスワード
          title: Password
          type: string
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          description: 体重（kg）
          title: Weight Kg
      required:
      - email
      - password
      - full_name
      title: UserCreate
      type: object
    UserLogin:
      description: ログインリクエスト
      properties:
        email:
          description: メールアドレス
          title: Email
          type: string
        password:
          description: パスワード
          title: Password
          type: string
      required:
      - email
      - password
      title: UserLogin
      type: object
    UserResponse:
      description: ユーザーレスポンス（パスワードなし）
      properties:
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        created_at:
          format: date-time
          title: Created At
          type: string
        email:
          title: Email
          type: string
        full_name:
          title: Full Name
          type: string
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
        height_cm:
          anyOf:
          - type: number
          - type: 'null'
          title: Height Cm
        id:
          title: Id
          type: string
        is_active:
          title: Is Active
          type: boolean
        is_verified:
          title: Is Verified
          type: boolean
        language:
          title: Language
          type: string
        last_login:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Login
        timezone:
          title: Timezone
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
        user_type:
          $ref: '#/components/schemas/UserType'
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Weight Kg
      required:
      - id
      - email
      - full_name
      - user_type
      - created_at
      - updated_at
      - is_active
      - is_verified
      - timezone
      - language
      title: UserResponse
      type: object
    UserStats:
      description: ユーザー統計情報
      properties:
        account_created_days:
          description: アカウント作成からの日数
          title: Account Created Days
          type: integer
        average_score:
          description: 平均スコア
          title: Average Score
          type: number
        latest_analysis_date:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: 最新分析日時
          title: Latest Analysis Date
        streak_days:
          default: 0
          description: 連続使用日数
          title: Streak Days
          type: integer
        this_month_analyses:
          description: 今月の分析回数
          title: This Month Analyses
          type: integer
        total_analyses:
          description: 総分析回数
          title: Total Analyses
          type: integer
      required:
      - total_analyses
      - this_month_analyses
      - average_score
      - account_created_days
      title: UserStats
      type: object
    UserType:
      description: ユーザータイプ
      enum:
      - standard
      - premium
      - professional
      - admin
      title: UserType
      type: string
    UserUpdate:
      description: ユーザー更新リクエスト
      properties:
        age:
          anyOf:
          - type: integer
          - type: 'null'
          description: 年齢
          title: Age
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          description: フルネーム
          title: Full Name
        gender:
          anyOf:
          - $ref: '#/components/schemas/Gender'
          - type: 'null'
          description: 性別
        height_cm:
          anyOf:
          - type: number
          - type: 'null'
          description: 身長（cm）
          title: Height Cm
        language:
          anyOf:
          - type: string
          - type: 'null'
          description: 言語設定
          title: Language
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          description: タイムゾーン
          title: Timezone
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          description: 体重（kg）
          title: Weight Kg
      title: UserUpdate
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    FirebaseAuth:
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            email: Email access
            openid: OpenID Connect
            profile: Profile access
      type: oauth2
info:
  contact:
    email: support@gaitanalysis.com
    name: Gait Analysis API Support
  description: "\n    ## Enhanced MediaPipe歩行分析API v2.0\n\n    スマートフォンで撮影した動画から歩行パターンを分析し、\n\
    \    理学療法士や医療従事者向けの詳細なレポートを生成する拡張APIです。\n\n    ### \U0001F680 新機能（v2.0 Enhanced）\n\
    \    - **30秒高速処理**: 撮影からレポート出力まで30秒以内\n    - **33ランドマーク活用**: MediaPipe Pose v0.10の全機能\n\
    \    - **代償動作検出**: 自動アラート機能付き\n    - **重心・バランス分析**: COM軌跡とバランス安定性評価\n    - **拡張出力**:\
    \ 正規化波形グラフ、スローモーション動画、カラーPDFレポート\n    - **iPad/iPhone最適化**: 60fps対応、横・後方撮影モード\n\
    \n    ### \U0001F4CA 詳細分析機能\n    - **スパイオテンポラル指標**: ピッチ、歩幅/ストライド長、立脚/遊脚時間、対称性指標\n\
    \    - **キネマティクス**: 股関節ROM、膝関節詳細角度、足関節背屈/底屈、骨盤・体幹回旋\n    - **代償動作アラート**: アウト/イントゥーイング、膝外反/内反、トレンデレンブルグ、ペルビックドロップ\n\
    \    - **バランス指標**: COM水平/垂直移動幅、重心投影線2Dマップ\n\n    ### \U0001F3AF 対象ユーザー\n    -\
    \ **理学療法士**: 詳細な歩行評価と治療計画立案\n    - **作業療法士**: 日常生活動作の評価\n    - **医師**: 診断支援と経過観察\n\
    \    - **研究者**: 歩行データの定量的解析\n    - **アスリート・トレーナー**: パフォーマンス向上\n\n    ### ⚙️ 技術仕様（Enhanced）\n\
    \    - **バックエンド**: FastAPI + Python 3.11+\n    - **AI処理**: MediaPipe Pose v0.10\
    \ + 最高精度モデル\n    - **並列処理**: マルチコア活用による高速化\n    - **データベース**: PostgreSQL + Redis\
    \ キャッシュ\n    - **認証**: Firebase Auth + JWT\n    - **監視**: リアルタイムパフォーマンス監視\n  \
    \  - **出力**: PDF/MP4/JSON/CSV 多形式対応\n    "
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Enhanced MediaPipe Gait Analysis API
  version: 2.0.0-enhanced
openapi: 3.1.0
paths:
...
  /api/v1/auth/firebase:
    post:
      description: "Firebase認証ログイン\n\nArgs:\n    firebase_token: FirebaseIDトークン\n\
        \    request: HTTPリクエスト\n    db: データベースセッション\n    \nReturns:\n    TokenResponse:\
        \ 認証トークンとユーザー情報"
      operationId: firebase_login_api_v1_auth_firebase_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HTTPAuthorizationCredentials'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Firebase Login
      tags:
      - auth
      - 認証
  /api/v1/auth/login:
    post:
      description: "ユーザーログイン\n\nArgs:\n    user_data: ログインデータ\n    request: HTTPリクエスト\n\
        \    db: データベースセッション\n    \nReturns:\n    TokenResponse: 認証トークンとユーザー情報"
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - auth
      - 認証
  /api/v1/auth/login-history:
    get:
      description: "ログイン履歴取得\n\nArgs:\n    limit: 取得件数\n    current_user: 現在のユーザー（認証情報）\n\
        \    db: データベースセッション\n    \nReturns:\n    dict: ログイン履歴"
      operationId: get_login_history_api_v1_auth_login_history_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Login History
      tags:
      - auth
      - 認証
  /api/v1/auth/me:
    get:
      description: "現在のユーザー情報取得\n\nArgs:\n    current_user: 現在のユーザー（認証情報）\n    db:\
        \ データベースセッション\n    \nReturns:\n    UserResponse: ユーザー情報"
      operationId: get_current_user_info_api_v1_auth_me_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Current User Info
      tags:
      - auth
      - 認証
    put:
      description: "現在のユーザー情報更新\n\nArgs:\n    user_data: 更新データ\n    current_user:\
        \ 現在のユーザー（認証情報）\n    db: データベースセッション\n    \nReturns:\n    UserResponse: 更新後のユーザー情報"
      operationId: update_current_user_api_v1_auth_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Update Current User
      tags:
      - auth
      - 認証
  /api/v1/auth/password/reset:
    post:
      description: "パスワードリセット要求\n\nArgs:\n    reset_data: パスワードリセットデータ\n    db: データベースセッション\n\
        \    \nReturns:\n    dict: リセット要求結果"
      operationId: request_password_reset_api_v1_auth_password_reset_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Request Password Reset
      tags:
      - auth
      - 認証
  /api/v1/auth/register:
    post:
      description: "新規ユーザー登録\n\nArgs:\n    user_data: ユーザー登録データ\n    request: HTTPリクエスト\n\
        \    db: データベースセッション\n    \nReturns:\n    TokenResponse: 認証トークンとユーザー情報"
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - auth
      - 認証
  /api/v1/auth/stats:
    get:
      description: "ユーザー統計情報取得\n\nArgs:\n    current_user: 現在のユーザー（認証情報）\n    db:\
        \ データベースセッション\n    \nReturns:\n    UserStats: ユーザー統計情報"
      operationId: get_user_statistics_api_v1_auth_stats_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get User Statistics
      tags:
      - auth
      - 認証
  /api/v1/auth/verify-token:
    post:
      description: "トークン検証\n\nArgs:\n    current_user: 現在のユーザー（認証情報）\n    \nReturns:\n\
        \    dict: トークン検証結果"
      operationId: verify_token_api_v1_auth_verify_token_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Verify Token
      tags:
      - auth
      - 認証
...
servers:
...